userprincipal 확인 필요

+++ src/main/frontend/src/components/Save.jsx
@@ -4,10 +4,10 @@ import { useHistory } from "react-router-dom";
 
 function Save(props) {
   const history = useHistory();
-  const item=props.item
-  const currentUser=props.currentUser;
-  console.log(currentUser)
-  const saveItem = async () => {
+  const item = props.item;
+  const currentUser = props.currentUser;
+
+  const saveItem = async (currentUser) => {
     try {
       const response = await axios.post(
         "/api/products",
@@ -18,7 +18,7 @@ function Save(props) {
           image: item.image,
           mallName: item.mallName,
           lprice: item.lprice,
-		currentUser
+          currentUser: currentUser
         },
         {
           headers: {
@@ -27,6 +27,7 @@ function Save(props) {
           },
         }
       );
+      alert("상품이 '찜한 품목'에 저장되었습니다!");
       return response.data.productId;
     } catch (error) {
       console.error(error);
@@ -35,18 +36,26 @@ function Save(props) {
     }
   };
 
-  const handleSaveButtonClick = async () => {
-    const savedItemId = await saveItem();
-    if (savedItemId) {
-      const savedItems = JSON.parse(localStorage.getItem('savedItems')) || [];
-      savedItems.push(savedItemId);
-      localStorage.setItem('savedItems', JSON.stringify(savedItems));
+  const onClickSave = async (e) => {
+    e.stopPropagation(); 
+    if (!props.authenticated) {
+       const confirmed = window.confirm("로그인이 필요한 서비스입니다. 로그인 페이지로 이동하시겠습니까?");
+      if (confirmed) {
+        history.push('/login');
+      }
+      
+    } else {
+        const savedItemId = await saveItem(currentUser);
+      if (savedItemId) {
+        const savedItems = JSON.parse(localStorage.getItem('savedItems')) || [];
+        savedItems.push(savedItemId);
+        localStorage.setItem('savedItems', JSON.stringify(savedItems));
+      }
     }
-  };
-
+  }
   return (
     <>
-      <button onClick={handleSaveButtonClick}>보관하기</button>
+      <button className='save-btn' onClick={onClickSave}>보관하기</button>
     </>
   );
 }
